+++ packages/misskey-js/src/autogen/types.ts
@@ -9503,6 +9503,7 @@ export interface operations {
                         manifestJsonOverride: string;
                         policies: Record<string, never>;
                         enableFanoutTimeline: boolean;
+                        fanoutTimelineActive: boolean;
                         enableFanoutTimelineDbFallback: boolean;
                         perLocalUserUserTimelineCacheMax: number;
                         perRemoteUserUserTimelineCacheMax: number;